backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
aiofiles
passlib[bcrypt]
python-jose[cryptography]
